chore(clean_up): update package names and reorganize test structure
- Renamed package `org.js.vento.webstormvento` to `org.js.vento.plugin` for consistency.
- Moved and refactored lexer and parser test cases for better organization.
- Added `VariableLexerTestCase` to test variable lexing scenarios.
