+++ src/test/kotlin/org/js/vento/plugin/lexer/CommentLexerTestCase.kt
@@ -3,7 +3,7 @@
  * All rights reserved.
  */
 
-package org.js.vento.webstormvento
+package org.js.vento.plugin.lexer
 
 class CommentLexerTestCase(name: String) : BaseLexerTestCase(name) {
     fun `test lexing comment`() {
